+++ extension/background/ng/services/lightning.js
@@ -122,7 +122,7 @@
 				case 'lnd_rest':
 					return new LNDClient(_analyticsService, $http, $window, options.restBaseUrl, options.macaroon, options.maxFeePercent);
 				case 'lnpayco':
-					return new LNPayClient(_analyticsService, $http, options.lnpaycoApiKey, options.lnpaycoWalletAccessKey);
+					return new LNPayClient(_analyticsService, $http, options.lnpaycoApiKey, options.lnpaycoWalletAccessKey, options.maxFeePercent);
 			}
 		}
 	}
@@ -164,7 +164,7 @@
 
 		sendPaymentWithKeySend(nodeId, amount, customRecordKey, customRecordValue, podcastPaymentMetadata) {
 			const that = this;
-			const feeLimit = this.calculateMaxFeeIn_mSats(amount);
+			const feeLimit = calculateMaxFeeIn_mSats(amount, this.maxFeeInPercent);
 			return this.buildPreimageAndPaymentHash().then((preimageAndPaymentHash) => {
 				const additionalCustomRecords = {};
 
@@ -229,10 +229,6 @@
 			});
 		}
 
-		calculateMaxFeeIn_mSats(amount) {
-			return Math.round(amount * this.maxFeeInPercent / 100);
-		}
-
 		buildPreimageAndPaymentHash() {
 			const randomValues = new ArrayBuffer(32);
 			const randomValuesUint8 = new Uint8Array(randomValues);
@@ -267,13 +263,14 @@
 	}
 
 	class LNPayClient {
-		constructor(_analyticsService, $http, apiKey, walletAccessKey) {
+		constructor(_analyticsService, $http, apiKey, walletAccessKey, maxFeeInPercent) {
 			this._analyticsService = _analyticsService;
 			this._$http = $http;
 			this._apiKey = apiKey;
 			this._walletAccessKey = walletAccessKey;
 			this._paymentQueue = [];
 			this._isSendingPayment = false;
+			this.maxFeeInPercent = maxFeeInPercent;
 		}
 
 		/**
@@ -328,9 +325,12 @@
 				return Promise.reject();
 			}
 
+			const feeLimit = calculateMaxFeeIn_mSats(amount, this.maxFeeInPercent);
+
 			let body = {
 				dest_pubkey: nodeId,
-				num_satoshis: LNPayClient.convertFrom_mSatsToSats(amount)
+				num_satoshis: LNPayClient.convertFrom_mSatsToSats(amount),
+				fee_limit_msat: feeLimit
 			}
 
 			if(customRecordKey) {
@@ -377,6 +377,10 @@
 		static convertFrom_mSatsToSats(amountIn_mSats) {
 			return Math.round(amountIn_mSats / 1000)
 		}
+	}
+
+	function calculateMaxFeeIn_mSats(amountInmSats, maxFeeInPercent) {
+		return Math.round(amountInmSats * maxFeeInPercent / 100);
 	}
 
 	function hexToString(hexstring) {